+++ packages/gafl-webapp-service/src/__tests__/server-fetch-public.js
@@ -1,11 +1,8 @@
 import { start, stop, initialize, injectWithoutSessionCookie } from '../__mocks__/test-utils-system.js'
 
-beforeAll(d => {
-  start(d)
-})
-
-beforeAll(d => initialize(d))
-afterAll(d => stop(d))
+beforeAll(() => new Promise(resolve => start(resolve)))
+beforeAll(() => new Promise(resolve => initialize(resolve)))
+afterAll(done => { stop(done) })
 
 describe('Where the server is started', () => {
   it('serve public resources without the session cookie', async () => {