+++ packages/gafl-webapp-service/src/__tests__/server-ga-integration.spec.js
@@ -28,7 +28,7 @@ jest.mock('../constants', () => ({
   CommonResults: {
     OK: 'okay'
   },
-  showDigitalLicencePages: {
+  ShowDigitalLicencePages: {
     YES: 'show-digital-licence-yes'
   }
 }))