+++ packages/gafl-webapp-service/src/pages/summary/licence-summary/result-function.js
@@ -1,12 +1,16 @@
-import { CONTACT_SUMMARY_SEEN, CommonResults, showDigitalLicencePages } from '../../../constants.js'
+import { CONTACT_SUMMARY_SEEN, CommonResults, ShowDigitalLicencePages } from '../../../constants.js'
 import { isPhysical } from '../../../processors/licence-type-display.js'
 
 export default async request => {
   const status = await request.cache().helpers.status.getCurrentPermission()
   const permission = await request.cache().helpers.transaction.getCurrentPermission()
 
-  if (status.renewal && isPhysical(permission) && permission.licensee.postalFulfilment !== false) {
-    return showDigitalLicencePages.YES
+  if (status.renewal) {
+    if (isPhysical(permission) && status.showDigitalLicencePages) {
+      return ShowDigitalLicencePages.YES
+    } else {
+      return CommonResults.SUMMARY
+    }
   }
 
   return status.fromSummary === CONTACT_SUMMARY_SEEN ? CommonResults.SUMMARY : CommonResults.OK