+++ packages/gafl-webapp-service/src/session-cache/__tests__/cache-manager.spec.js
@@ -1,9 +1,9 @@
 import { start, stop, initialize, injectWithCookies, mockSalesApi } from '../../__mocks__/test-utils-system.js'
 import { CONTROLLER, LICENCE_TYPE } from '../../uri.js'
 
-beforeAll(d => start(d))
-beforeAll(d => initialize(d))
-afterAll(d => stop(d))
+beforeAll(() => new Promise(resolve => start(resolve)))
+beforeAll(() => new Promise(resolve => initialize(resolve)))
+afterAll((d) => stop(d))
 
 mockSalesApi()